PRE-COMMIT: add sample_metadata/ to exclusions in pre-commit file as per https://stackoverflow.com/a/61046953
